requirements.txt
cassandra-driver==3.7.0
click==6.6
futures==3.0.5
invoke==0.13.0
six==1.10.0
